+++ projects/ngx-googlemaps-tracking-view/src/lib/components/geolocation-button/geolocation-button.ts
@@ -1,5 +1,5 @@
 /// <reference types="@types/googlemaps" />
-import { Component, Output, EventEmitter } from '@angular/core';
+import { Component, Output, EventEmitter, Input } from '@angular/core';
 import geolocationMarkrFn from 'geolocation-marker';
 import { GoogleMapsWrapper } from '../../services/googlemaps-wrapper';
 geolocationMarkrFn;
@@ -33,6 +33,11 @@ declare class GeolocationMarker {
 })
 export class GeolocationButtonComponent {
   /**
+   * If `true`, the button won't attempt to center the map on the user when clicked,
+   * but will emit the event `click` instead.
+   */
+  @Input() useCustomClickFn = false;
+  /**
    * Emited after obtaining the user's location
    */
   @Output() locate = new EventEmitter<google.maps.LatLng>();
@@ -41,24 +46,38 @@ export class GeolocationButtonComponent {
   constructor(protected googlemapsWrapper: GoogleMapsWrapper) { }
 
   /**
-   * Center map on user's location
+   * Get user's location and center map on it
    */
   centerOnUser() {
-    if (navigator.geolocation) {
-      navigator.geolocation.getCurrentPosition((position) => {
-        const lng = position.coords.longitude;
-        const lat = position.coords.latitude;
-        this.googlemapsWrapper.map.setCenter({ lat, lng });
+    if (this.useCustomClickFn) {
+      return;
+    }
 
-        if (this._geoLocationMarker) this._geoLocationMarker.setMap(null);
-        this._geoLocationMarker = new GeolocationMarker(this.googlemapsWrapper.map);
-        this.locate.emit(this._geoLocationMarker.position);
-      }, error => {
-        console.error(error);
-      },
-        { enableHighAccuracy: true });
+    if (navigator.geolocation) {
+      navigator.geolocation.getCurrentPosition(this.onPosition, console.error, { enableHighAccuracy: true });
     } else {
       console.error('No support for geolocation');
     }
+  }
+
+  /**
+   * Center the map on a specific position and draw the geoLocationMarker
+   */
+  setUserPosition(lat: number, lng: number) {
+    this.googlemapsWrapper.map.setCenter({ lat, lng });
+    if (this._geoLocationMarker) this._geoLocationMarker.setMap(null);
+    this._geoLocationMarker = new GeolocationMarker(this.googlemapsWrapper.map);
+    this.locate.emit(this._geoLocationMarker.position);
+  }
+
+  private onPosition(position: Position) {
+    const lng = position.coords.longitude;
+    const lat = position.coords.latitude;
+    // this.googlemapsWrapper.map.setCenter({ lat, lng });
+
+    // if (this._geoLocationMarker) this._geoLocationMarker.setMap(null);
+    // this._geoLocationMarker = new GeolocationMarker(this.googlemapsWrapper.map);
+    // this.locate.emit(this._geoLocationMarker.position);
+    this.setUserPosition(lat, lng);
   }
 }